grafana v0.0.1 Grafana.Currentorg

Functions

add_user(json)

Add a user to the current organization, as specified by json.

delete_user(id)

Delete an existing user with given id.

get()

Get the current organization.

iex> {:ok, org} = Grafana.Currentorg.get …> Map.keys(org) [“address”, “id”, “name”]

get_users()

Get all users within the current organization.

iex> {:ok, org_users} = Grafana.Currentorg.get_users …> Map.keys(hd(org_users)) [“email”, “login”, “orgId”, “role”, “userId”]

update(json)

Update the current organization as specified by json.

update_user(id, json)

Update an existing user with given id, as specified by json.