grafana v0.1.0 Grafana.Currentuser

Functions

change_password(json)

Change password for the current user.

get()

Get the actual current user.

iex> {:ok, user} = Grafana.Currentuser.get …> Map.keys(user) [“email”, “isGrafanaAdmin”, “login”, “name”, “orgId”, “theme”]

orgs()

Get orgs for the actual current user.

iex> {:ok, orgs} = Grafana.Currentuser.orgs …> Map.keys(hd(orgs)) [“name”, “orgId”, “role”]

star_dashboard(dashboard_id)

Star dashboard with given id for the current user.

switch_context(org_id)

Change organization context for the current user.

unstar_dashboard(dashboard_id)

Unstar dashboard with given id for the current user.