grafana v0.1.1 Grafana.Dashboard

Functions

delete(slug)

Delete an existing dashboard with the given slug. Slug is the url-friendly version of the dashboard title.

iex> {:ok, response} = Grafana.Dashboard.delete(“test”) …> Map.keys(response) [“title”]

get()

get(slug)

Get a dashboard with the given slug. Slug is the url-friendly version of the dashboard title.

iex> {:ok, dash} = Grafana.Dashboard.get “worldping-endpoint-dns” …> Map.keys(dash) [“dashboard”, “meta”]

get_from_json(json)

Get dashboard from json file.

home()

Get the home dashboard

iex> {:ok, dash} = Grafana.Dashboard.home …> Map.keys(dash) [“dashboard”, “meta”]

iex> {:ok, dash} = Grafana.Dashboard.get …> Map.keys(dash) [“dashboard”, “meta”]

new(json)

Create a new dashboard specified by json.

params_to_json(id \\ "null", title \\ "new dashboard", tags \\ [], rows \\ [], overwrite \\ false, version \\ 0, editable \\ false, style \\ "light")

Format dashboard parameters into JSON.

Model: { “dashboard”: {

"id": null,
"title": "Production Overview",
"tags": [ "templated" ],
"timezone": "browser",
"rows": [
  {
  }
],
"schemaVersion": 6,
"version": 0

}, “overwrite”: false }

search(query, starred, tag, tagcloud)

Search dashboards with given parameters. “starred” is a boolean flag indicating whether only starred dashboards should be returned. “tagcloud” is a boolean indicating if a tagcloud should be returned.

iex> {:ok, search} = Grafana.Dashboard.search(“”, false, “Endpoints”, false) …> Map.keys(hd(search)) [“id”, “isStarred”, “tags”, “title”, “type”, “uri”]

tags()

Get all the available tags for dashboards.

iex> {:ok, tags} = Grafana.Dashboard.tags …> Map.keys(hd(tags)) [“count”, “term”]

update(json)

Update an existing dashboard as specified in json.

iex> {:ok, dash} = Grafana.Dashboard.update(%{}) …> Map.keys(dash) [“slug”, “status”, “version”]