grafana v0.1.3 Grafana.Organizations

Functions

add_user(org_id, json)

Add user to given organization, as specified by json.

create(json)

Create an organization

delete_user(org_id, user_id)

Delete given user inside given organization.

get()

Get all organizations.

get(id)

Get a specific organization with given id.

get_users(org_id)

Get all users within a given organization.

org_params_to_json(name)

update(id, json)

Update organization with id, as specified by json.

update_user(org_id, user_id, json)

Update given user inside given organization, as specified by json.